requirements.api.txt
fastapi
uvicorn
google-cloud-firestore
requests
